refactor: keep the traffic panel's module to components
Fast refresh only survives a module that exports components alone, so
the icon map, the source labels and the arrivals hook move to a module
of their own and the Waiting/Confirmed marker becomes a component
rather than a function returning one.
