+++ client/dashboard/src/pages/setup/components/traffic-activity.ts
@@ -1,0 +1,110 @@
+import { useEffect, useRef, useState } from "react";
+
+const MAX_EVENTS_SHOWN = 8;
+// Spacing between visible event arrivals. When a poll returns multiple new
+// events, we queue them and play one in every PLAYBACK_INTERVAL_MS so the
+// sliding-window animation reads cleanly (1 in, 1 out) instead of a cascade.
+const PLAYBACK_INTERVAL_MS = 400;
+
+export const SOURCE_ICONS: Record<string, string> = {
+  "claude-code": "/icons/platforms/claude.svg",
+  "claude-code-desktop": "/icons/platforms/claude.svg",
+  // Inference hook deliveries: Claude conversations on claude.ai and Claude
+  // Code's web surface, both reported by Anthropic rather than by a plugin.
+  "claude-chat-web": "/icons/platforms/claude.svg",
+  "claude-code-web": "/icons/platforms/claude.svg",
+  cowork: "/icons/platforms/claude.svg",
+  cursor: "/icons/platforms/cursor.svg",
+  codex: "/icons/platforms/openai.svg",
+  chatgpt: "/icons/platforms/openai.svg",
+  "chatgpt-work": "/icons/platforms/openai.svg",
+};
+
+export function sourceLabel(source: string): string {
+  switch (source) {
+    case "claude-code":
+      return "Claude Code";
+    case "claude-code-desktop":
+      return "Claude Desktop";
+    case "claude-chat-web":
+      return "Claude Chat Web";
+    case "claude-code-web":
+      return "Claude Code Web";
+    case "cursor":
+      return "Cursor";
+    case "codex":
+      return "Codex";
+    case "chatgpt":
+      return "ChatGPT";
+    case "chatgpt-work":
+      return "ChatGPT Work";
+    case "cowork":
+    case "claude-cowork":
+      return "Cowork";
+    default:
+      return source;
+  }
+}
+
+/**
+ * One line in the live tail, whatever produced it. Hook events and inference
+ * hook conversations reach the panel through different reads, so they are
+ * flattened to this before it sees them.
+ */
+export interface TrafficActivity {
+  /** Stable identity, so a repeated poll doesn't replay the same row. */
+  key: string;
+  /** Product surface, for the icon and the fallback label. */
+  source: string;
+  /** Who produced it, e.g. an email. Falls back to the source's label. */
+  actor?: string;
+  /** What happened, e.g. "Tool call: Bash". */
+  action: string;
+  timeMs: number;
+}
+
+/**
+ * Accumulates whatever the caller's poll turned up into the visible window,
+ * one arrival at a time. Activities already seen are ignored, so a feed that
+ * re-reports the same rows (a listing rather than a cursor) is safe to pass in
+ * whole on every tick.
+ */
+export function useTrafficArrivals(incoming: TrafficActivity[]): {
+  events: TrafficActivity[];
+  hasEvents: boolean;
+} {
+  const [events, setEvents] = useState<TrafficActivity[]>([]);
+  const [totalReceived, setTotalReceived] = useState(0);
+  // Pending playback queue — new activities land here and drain into `events`
+  // one at a time via the playback interval.
+  const queueRef = useRef<TrafficActivity[]>([]);
+  const seenKeysRef = useRef<Set<string>>(new Set());
+
+  // Drain the playback queue one event at a time so the sliding window
+  // animation can be observed per arrival, even when bursts deliver many at
+  // once.
+  useEffect(() => {
+    const id = setInterval(() => {
+      if (queueRef.current.length === 0) return;
+      const next = queueRef.current.shift();
+      if (!next) return;
+      setEvents((prev) => [next, ...prev].slice(0, MAX_EVENTS_SHOWN));
+    }, PLAYBACK_INTERVAL_MS);
+    return () => clearInterval(id);
+  }, []);
+
+  useEffect(() => {
+    const fresh = incoming.filter((activity) => {
+      if (seenKeysRef.current.has(activity.key)) return false;
+      seenKeysRef.current.add(activity.key);
+      return true;
+    });
+    if (fresh.length === 0) return;
+    // Feeds return newest-first; reverse so we enqueue oldest-first and the
+    // newest event still ends up at the top of the visible stack.
+    queueRef.current.push(...[...fresh].reverse());
+    setTotalReceived((prev) => prev + fresh.length);
+  }, [incoming]);
+
+  return { events, hasEvents: totalReceived > 0 };
+}
